Moved pll calc into separate file
