+++ himbaechel/uarch/gatemate/pack.h
@@ -42,6 +42,7 @@ struct GateMatePacker
     void remove_constants();
 
     void pll_out(CellInfo *cell, IdString origPort, int placement);
+    PllCfgRecord get_pll_settings(double f_ref, double f_core, int mode, int low_jitter, bool pdiv0_mux, bool feedback);
 
     CellInfo *move_ram_i(CellInfo *cell, IdString origPort, int placement);
     CellInfo *move_ram_o(CellInfo *cell, IdString origPort, int placement);
